[Movie] Move movie class to data folder
