+++ data/ImageProviderInterface.h
@@ -6,7 +6,7 @@
 #include "data/Concert.h"
 #include "data/TvShow.h"
 #include "globals/Globals.h"
-#include "movies/Movie.h"
+#include "data/Movie.h"
 
 class TvShow;
