+++ data/Storage.h
@@ -10,7 +10,7 @@
 #include "data/TvShowEpisode.h"
 #include "export/ExportTemplate.h"
 #include "globals/Globals.h"
-#include "movies/Movie.h"
+#include "data/Movie.h"
 #include "music/Album.h"
 #include "music/Artist.h"
