+++ export/ExportDialog.h
@@ -8,7 +8,7 @@
 #include "data/TvShow.h"
 #include "data/TvShowEpisode.h"
 #include "export/ExportTemplate.h"
-#include "movies/Movie.h"
+#include "data/Movie.h"
 
 namespace Ui {
 class ExportDialog;
