+++ globals/ImageDialog.h
@@ -15,7 +15,7 @@
 #include "data/TvShow.h"
 #include "data/TvShowEpisode.h"
 #include "globals/Globals.h"
-#include "movies/Movie.h"
+#include "data/Movie.h"
 #include "music/Album.h"
 #include "music/Artist.h"
 #include "smallWidgets/ImageLabel.h"
