+++ globals/TrailerDialog.h
@@ -10,7 +10,7 @@
 #include <QVideoWidget>
 
 #include "globals/Globals.h"
-#include "movies/Movie.h"
+#include "data/Movie.h"
 
 namespace Ui {
 class TrailerDialog;
