+++ renamer/Renamer.h
@@ -8,7 +8,7 @@
 #include "data/Concert.h"
 #include "data/TvShow.h"
 #include "data/TvShowEpisode.h"
-#include "movies/Movie.h"
+#include "data/Movie.h"
 
 namespace Ui {
 class Renamer;
